chore: connecting redux to the modal
